+++ client/src/components/GoogleOAuthButton/index.tsx
@@ -1,27 +1,14 @@
 import React from "react";
-import { useNavigate } from "react-router-dom";
-import { useMutation } from "react-query";
+import useGoogleLogin from "@hooks/query/useGoogleLogin";
 import { GoogleLogin, CredentialResponse } from "@react-oauth/google";
-import { RoutePath } from "@constants/enums";
-import UserAPI from "@api/userAPI";
 
 const GoogleOAuthButton = () => {
-  const navigate = useNavigate();
-  const { mutate } = useMutation((accessToken: string) => UserAPI.googleLogin(accessToken), {
-    onSuccess: (isExistingUser) => {
-      console.log(isExistingUser);
-      if (isExistingUser) {
-        return;
-      }
-      navigate(RoutePath.JOIN);
-    },
-  });
+  const { googleLogin } = useGoogleLogin();
 
   const handleSuccessLogin = ({ credential }: CredentialResponse) => {
-    if (!credential) {
-      return;
+    if (credential) {
+      googleLogin({ accessToken: credential });
     }
-    mutate(credential);
   };
 
   return <GoogleLogin onSuccess={handleSuccessLogin} useOneTap />;
